backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
psycopg2-binary
kafka-python
pika
python-dotenv
